+++ sys/decklink/gstdecklinkvideosrc.cpp
@@ -423,7 +423,8 @@ gst_decklink_video_src_set_caps (GstBaseSrc * bsrc, GstCaps * caps)
         gst_decklink_get_connection (self->connection));
     if (ret != S_OK) {
       GST_ERROR_OBJECT (self,
-          "Failed to set configuration (input source): 0x%08lx", ret);
+          "Failed to set configuration (input source): 0x%08lx",
+          (unsigned long) ret);
       return FALSE;
     }
 
@@ -432,7 +433,8 @@ gst_decklink_video_src_set_caps (GstBaseSrc * bsrc, GstCaps * caps)
           bmdAnalogVideoFlagCompositeSetup75);
       if (ret != S_OK) {
         GST_ERROR_OBJECT (self,
-            "Failed to set configuration (composite setup): 0x%08lx", ret);
+            "Failed to set configuration (composite setup): 0x%08lx",
+            (unsigned long) ret);
         return FALSE;
       }
     }
@@ -449,7 +451,8 @@ gst_decklink_video_src_set_caps (GstBaseSrc * bsrc, GstCaps * caps)
           &autoDetection);
       if (ret != S_OK) {
         GST_ERROR_OBJECT (self,
-            "Failed to get attribute (autodetection): 0x%08lx", ret);
+            "Failed to get attribute (autodetection): 0x%08lx",
+            (unsigned long) ret);
         return FALSE;
       }
       if (autoDetection)
@@ -467,7 +470,8 @@ gst_decklink_video_src_set_caps (GstBaseSrc * bsrc, GstCaps * caps)
   format = self->caps_format;
   ret = self->input->input->EnableVideoInput (mode->mode, format, flags);
   if (ret != S_OK) {
-    GST_WARNING_OBJECT (self, "Failed to enable video input: 0x%08lx", ret);
+    GST_WARNING_OBJECT (self, "Failed to enable video input: 0x%08lx",
+        (unsigned long) ret);
     return FALSE;
   }
 
@@ -699,7 +703,7 @@ gst_decklink_video_src_got_frame (GstElement * element,
       res = dtc->GetComponents (&hours, &minutes, &seconds, &frames);
       if (res != S_OK) {
         GST_ERROR ("Could not get components for timecode %p: 0x%08lx", dtc,
-            res);
+            (unsigned long) res);
         f.tc = NULL;
       } else {
         bflags = dtc->GetFlags ();
@@ -1031,7 +1035,7 @@ gst_decklink_video_src_start_streams (GstElement * element)
     res = self->input->input->StartStreams ();
     if (res != S_OK) {
       GST_ELEMENT_ERROR (self, STREAM, FAILED,
-          (NULL), ("Failed to start streams: 0x%08lx", res));
+          (NULL), ("Failed to start streams: 0x%08lx", (unsigned long) res));
       return;
     }
   } else {
@@ -1085,7 +1089,7 @@ gst_decklink_video_src_change_state (GstElement * element,
       res = self->input->input->StopStreams ();
       if (res != S_OK) {
         GST_ELEMENT_ERROR (self, STREAM, FAILED,
-            (NULL), ("Failed to stop streams: 0x%08lx", res));
+            (NULL), ("Failed to stop streams: 0x%08lx", (unsigned long) res));
         ret = GST_STATE_CHANGE_FAILURE;
       }
       break;
